requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
openai>=1.20.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
